Do not rely on local file system for kubeflow orchestration engine.
tfx CLI stores pipeline information on the local storage and it created many usability issues (#2420) because there are two copies of pipeline information, one on the local and one in the Kubeflow Pipelines cluster.
Changes tfx CLI to use the information in the Kubeflow Pipelines clusters only, and do not store any information on the disk.

PiperOrigin-RevId: 348890570